added navbar and routing, moved home page to seperate component
