+++ ui/package.json
@@ -15,6 +15,7 @@
     "react-datepicker": "^8.2.1",
     "react-dom": "^19.0.0",
     "react-router": "^7.4.1",
+    "react-router-bootstrap": "^0.26.3",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
