+++ ui/src/Pages/HomePage.css
@@ -1,0 +1,83 @@
+.App {
+    display: flex;
+    justify-content: center;
+    height: 100vh;
+    background-color: #f8f9fa;
+    padding: 20px;
+}
+
+.birdie-button-column {
+    display: flex;
+    flex-direction: column;
+    gap: 15px;
+    align-items: center;
+    padding: 8px;
+}
+
+
+
+.birdie-button-wrapper {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 8px;
+    padding: 17px;
+    background: white;
+    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+    border-radius: 12px;
+    cursor: pointer;
+}
+
+.birdie-button {
+    background: none;
+    border: none;
+    padding: 16px;
+}
+
+.birdie-image {
+    width: 60px;
+    height: 60px;
+}
+
+.session-card {
+    background: white;
+    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+    border-radius: 12px;
+    padding: 20px;
+    width: 100%;
+    max-width: 500px;
+}
+
+.session-title {
+    font-size: 22px;
+    font-weight: bold;
+    color: #333;
+}
+
+.session-date {
+    font-size: 14px;
+    color: #666;
+}
+
+.session-info {
+    font-size: 16px;
+    font-weight: 500;
+    margin-bottom: 8px;
+}
+
+.unpaid-title {
+    font-size: 16px;
+    font-weight: 600;
+    color: #d32f2f;
+}
+
+.unpaid-player {
+    font-size: 14px;
+    color: #444;
+}
+
+.no-unpaid {
+    font-size: 14px;
+    color: #888;
+}
+
